Split packaging test into separate file
